combined approuter and app js files
